+++ packages/@expo/metro-runtime/src/async-require/loadBundle.ts
@@ -20,10 +20,9 @@ export async function loadBundleAsync(bundlePath: string): Promise<void> {
   if (process.env.NODE_ENV === 'production') {
     return fetchThenEvalAsync(requestUrl);
   } else {
-    const Platform = require('react-native').Platform;
     const LoadingView = require('../LoadingView')
       .default as typeof import('../LoadingView').default;
-    if (Platform.OS !== 'web') {
+    if (process.env.EXPO_OS !== 'web') {
       // Send a signal to the `expo` package to show the loading indicator.
       LoadingView.showMessage('Downloading...', 'load');
     }
@@ -35,7 +34,7 @@ export async function loadBundleAsync(bundlePath: string): Promise<void> {
         HMRClient.registerBundle(requestUrl);
       })
       .finally(() => {
-        if (!--pendingRequests && Platform.OS !== 'web') {
+        if (!--pendingRequests && process.env.EXPO_OS !== 'web') {
           LoadingView.hide();
         }
       });